+++ Enterprise/Subsystem/ProductLauncher/Database/Identity/Person/Stored Procedures/ListPersons_Ver04.sql
@@ -208,7 +208,7 @@ BEGIN
   INSERT INTO @tblPrimaryProperties (        
    Properties        
   )        
-  SELECT CONVERT(int, value)        
+  SELECT CONVERT(bigint, value)        
   FROM STRING_SPLIT(@PrimaryProperties, ',');        
  END
